requirements.txt
opencv-python
numpy
face_recognition
pillow
requests
fastapi
uvicorn